+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1.115] - 2026-02-10
+
+### Summary
+
+feat(tests): configuration management system
+
+### Test
+
+- update tests/test_builders.py
+
+
 ## [0.1.114] - 2026-02-10
 
 ### Summary